[EPT-1715_NEW_PAGE_MODAL_1] new component for create page modal
